+++ put_numbers.c
@@ -6,7 +6,7 @@
 /*   By: jthuy <jthuy@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/22 12:31:02 by jthuy             #+#    #+#             */
-/*   Updated: 2019/12/23 17:46:38 by jthuy            ###   ########.fr       */
+/*   Updated: 2019/12/23 18:46:27 by jthuy            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,7 @@ void	put_sign(void **value, unsigned int *flags, int *amount)
 	}
 }
 
-void	put_sign_f(double *value_f, unsigned int *flags, int *amount)
+void	put_sign_f(long double *value_f, unsigned int *flags, int *amount)
 {
 	unsigned long	test;
 	ft_memcpy(&test, &(*value_f), 8);
